Update README.md
Another one to my collection of front-end web games! This game is extremely basic, and you just press any key to start the game which triggers the game mechanism that beeps one of the four colored boxes on the screen. The beeped box goes into an array which stores the sequence of color buttons to be pressed. The user then has to copy that particular sequence and once the full sequence has been copied another colored button gets added to that sequence making the sequence longer as each level progresses. The user can make upto as many stages as they are able to